backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
nepali-datetime==1.0.8.4
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
